pyproject.toml
[project]
name = "meu-projeto"
version = "0.1.0"
description = ""
authors = [
    {name = "kadu",email = "kadukaua@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.135.3,<0.136.0)",
    "uvicorn[standart] (>=0.44.0,<0.45.0)",
    "sqlmodel (>=0.0.38,<0.0.39)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "bcrypt (==4.0.1)"
]

[tool.poetry]
packages = [{include = "meu_projeto", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
